+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # module name - release notes
 
 ## v1.0.2
+- Enabled Controller security group management
 - Fix issue where NSG throws an error getting a null in stead of a string, when disabling deployment of copilot.
 
 ## v1.0.1